vessel_segmentation_3d/requirements.txt
# 血管分割与三维重建系统依赖包
# =====================================

# 核心依赖
# -------
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 深度学习框架
# ------------
torch>=1.10.0
torchvision>=0.11.0

# 医学影像处理
# ------------
SimpleITK>=2.1.0
nibabel>=3.2.0
pydicom>=2.2.0
pyradiomics>=3.0.0

# 图像处理
# --------
scikit-image>=0.18.0
opencv-python>=4.5.0
Pillow>=8.3.0

# 图论和网络分析
# --------------
networkx>=2.6.0

# 可视化
# ------
matplotlib>=3.4.0
seaborn>=0.11.0
pyvista>=0.32.0
mayavi>=4.7.0

# MONAI - 医学影像AI框架
# ---------------------
monai>=1.0.0

# 数据处理
# --------
tqdm>=4.62.0
PyYAML>=5.4.0

# 可选依赖
# --------
# 3D可视化
# vtk>=9.1.0

# 高级骨架化
# itk>=5.2.0

# 机器学习
# scikit-learn>=1.0.0
# xgboost>=1.5.0

# 开发工具
# --------
# pytest>=6.2.0
# black>=21.7b0
# flake8>=3.9.0

# 文档
# ---
# sphinx>=4.2.0
# sphinx-rtd-theme>=1.0.0

# 安装说明
# ========
# 基础安装：
#   pip install -r requirements.txt
#
# 完整安装（包含可选依赖）：
#   pip install -r requirements.txt --extras all
#
# 开发安装：
#   pip install -r requirements.txt --extras dev
